Move url to common, added section in readme
