+++ jeeves/jobs.py
@@ -14,6 +14,18 @@ def get_stage_failure(build_stages):
 	return 'N/A'
 
 
+def generate_failure_stage_log_urls(config, err_stage, job_url, lcb_num):
+	''' generates list of urls for failed build stages
+	'''
+	stage_urls = []
+	if 'stage_logs' in config and err_stage in config['stage_logs']:
+		for path in config['stage_logs'][err_stage]:
+			stage_urls.append("{}/{}/artifact/{}".format(job_url, lcb_num, path))
+	if not stage_urls:
+		stage_urls = None
+	return stage_urls
+
+
 def get_jenkins_job_info(server, job_name, filter_param_name=None, filter_param_value=None):
 	''' takes in jenkins server object and job name
 		optionally takes name and value of jenkins param to filter builds by
